+++ tests/AllTests.php
@@ -4,6 +4,7 @@ namespace XynhaCS\Tests;
 
 use PHPUnit\Framework\TestSuite;
 
+use XynhaCS\Tests\Sniffs\Arrays\TsArray;
 use XynhaCS\Tests\Sniffs\WhiteSpace\TsWhiteSpace;
 
 final class AllTests
@@ -15,6 +16,9 @@ final class AllTests
 
         $ws = new TsWhiteSpace();
         $suite->addTest($ws->getTestSuite());
+
+        $array = new TsArray();
+        $suite->addTest($array->getTestSuite());
 
         return $suite;
     }